Update bbox and kwargs strategy
The bbox for passing limits to plots is renamed to limits because it was
too similar to the bbox_inches argument used by matplotlib for a totally
different concept.

limits and vaac_colours were made explicit arguments for plot_2d_cube,
however if these keywords are also present within a **kwargs that is
passed to the function, a ValueError is raised.  So plot_2d_cube no
longer accepts **kwargs, and the upacking of limits and vaac_colours
takes place in the plot_3d_cube and plot_4d_cube functions before they
call it.