Add CF Standard Names to derived cubes
The cubes created by Hysplit are all derived from a single variable,
whose name is not standard, and so must be given names when created.  We
use names based on the NetCDF standard list:

http://cfconventions.org/Data/cf-standard-names/71/build/cf-standard-name-table.html

The cubes for NAME and Fall3D are not renamed as is makes it harder to
relate to original model output.  An additional attribute, CF Standard
Name is added to provide a source of a consistent name between model types.