Use lru_cache for properties
The lru_cache decorator stores the results when the
function is called, and returns the stored values whenever it is called
with the same arguments in the future.  When it is applied to properties (e.g.
total_deposition), it saves on calculations each time that they are accessed.  This particularly applies to Hysplit, where properties must be derived via calculation.  Manual testing shows that this didn't make much difference to speed of execution, but it did reduce the number of units warnings displayed as each property was only calculated once.