Extract _save_yx_slice_figure from plot_3d_cube
Here we attempt to extract the repeated code from plot_3d_cube and
plot_4d_cube.  We need to create a dictionary that is shared between
processes to allow metadata information to be collected.

This code doesn't work because the multi_plot_yx_slice function is not
defined on the top level of the module and so cannot be pickled to be
used with multiprocessing.