Set matplotlib.backend to 'Agg'
The Agg backend is required for plotting without a display. Setting it
in the plotting code results in a warning as it must be set before
matplotlib.pyplot is imported, and that happens when setting up Iris
imports.  Setting in ash_model_plotting/__init__.py ensures that it is
set before any other imports.