Use spawn to create new processes for Pool
Adding the multiprocessing Pool caused processing to hang in some
situations.  The cause was likely to be due to matplotlib keeping hold
of references to figures, or perhaps some other globally imported
module objects.  The problem was shown most commonly when running the
tests, suggesting that pytest may also have been a factor.

This commit uses `spawn` to create each process in the Pool, as opposed
to the default (on Linux) `fork` method.  Spawn ensures that the child
processes only inherit essential resources and are better separated from
the parent.  It seems to have done the trick.

For some background, see https://pythonspeed.com/articles/python-multiprocessing/