Realize cube data to avoid lazy loading
Iris cubes have lazy loading, where data are not loaded until needed.
Calling convert_units() on a cube with lazy data defines a function
within a function that will be used to convert the data when they are
loaded.  Unfortunately, this function cannot be "pickled" in order to
send to each process of the multiprocessing Pool.

This commit calls cube.data to load cube data before it is passed for
multiprocessing.  This avoids the issues with lazy loading.